lorabridge/requirements.txt
python-telegram-bot>=20.7
meshtastic>=2.2.13
python-dotenv>=1.0.1
rich>=13.7.0
